충돌수정
